requirements.txt
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Faker==30.8.1
grpcio==1.66.2
grpcio-tools==1.66.2
h11==0.14.0
idna==3.10
mailosaur==7.18.0
natsort==8.4.0
outcome==1.3.0.post0
overrides==7.7.0
protobuf==5.27.2
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
robotframework==7.1.1
robotframework-assertion-engine==3.0.3 
robotframework-browser==18.9.1
robotframework-mailclient==0.0.12
robotframework-pythonlibcore==4.4.1
robotframework-retryfailed==0.2.0
seedir==0.5.0
selenium==4.25.0
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
websocket-client==1.8.0
wrapt==1.16.0
wsproto==1.2.0